services/olap-studio/pyproject.toml
[project]
name = "olap-studio"
version = "1.0.0"
description = "Axiom OLAP Studio — 스타 스키마 OLAP 피벗 · 큐브 관리 · ETL · 리니지"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "asyncpg>=0.30.0",
    "python-jose[cryptography]>=3.3.0",
    "structlog>=24.0.0",
    "lxml>=5.1.0",
    "httpx>=0.27.0",
    "redis>=5.0.0",
    "openai>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]
ai = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
]
